Move ur_moveit launch into moveit_config package.
