+++ ur_bringup/CMakeLists.txt
@@ -9,8 +9,4 @@ install(DIRECTORY config launch
   DESTINATION share/${PROJECT_NAME}
 )
 
-# Install Python modules
-ament_python_install_package(${PROJECT_NAME})
-ament_python_install_module(ur_bringup/launch_common.py)
-
 ament_package()
