+++ ur_moveit_config/CMakeLists.txt
@@ -2,9 +2,15 @@ cmake_minimum_required(VERSION 3.5)
 project(ur_moveit_config)
 
 find_package(ament_cmake REQUIRED)
+find_package(ament_cmake_python REQUIRED)
+find_package(rclpy REQUIRED)
 
-install(DIRECTORY config rviz srdf
+install(DIRECTORY config launch rviz srdf
   DESTINATION share/${PROJECT_NAME}
 )
+
+# Install Python modules
+ament_python_install_package(${PROJECT_NAME})
+ament_python_install_module(${PROJECT_NAME}/launch_common.py)
 
 ament_package()
